+++ test/spellcheck_test.py
@@ -13,7 +13,7 @@ def test_spellcheck_generator():
 
 def test_spellcheck_unicode():
     path = helper.fixture('unicode.txt')
-    check = spellcheck.NorvigSpellCheck(cracklib=path)
+    check = spellcheck.NorvigSpellCheck(words=path)
     assert_spellcheck(check)